refactor: 클래스 파일의 패키지 변경
BridgeGame: /bridge -> /bridge/model/domain
InputView: /bridge -> /bridge/view
OutputView /bridge -> /bridge/view
