+++ src/main/java/bridge/model/domain/BridgeGame.java
@@ -1,4 +1,4 @@
-package bridge;
+package bridge.model.domain;
 
 /**
  * 다리 건너기 게임을 관리하는 클래스
